refactor: SplashPage에 있는 내용을 컴포넌트로 분리
- metadata 작성을 위해 이와 같이 구현
